vendorMgmt/requirements.txt
django==5.0.4
djangorestframework==3.15.1
faker==25.0.1
djangorestframework-simplejwt==5.3.1
